refactor(dashboard): extract tunnel-ready polling into useTunnelReady
Pure move of the tunnel-readiness state + polling effect out of App.tsx
into a dedicated hook. Same deps ([isConnected]), same dynamic imports,
same 3s retry cadence — no behavior change.

Part of #5560 (App.tsx Shell decomposition).
